unused imports deleted var changed to const somewhere CustomCard improved structure changes
