+++ src/components/CustomCard/assets/cow.png
