+++ src/components/StatQuote/assets/brat.jpg
