+++ src/components/StatQuote/assets/cowboy.jpg
