+++ src/components/StatQuote/assets/samurai.jpg
