+++ src/components/StatQuote/assets/wolf.jpg
