+++ src/components/StatQuote/index.css
